Use octet each for class and method id (changed c++ and python)
Modified indexes in xml for message.empty, message.offset and the c++ cluster class

Fixed encoding for rfc1982-long-set in c++ and python (its a size not a count that is prepended)

Fixed minor typo in configuration option help string

Use session.open/close in python tests, handle session.closed

Commented out the response tag in session.close due to pythons ambiguity as to whether session.closed is a response or not

Disabled broker.test_closed_channel (due to above issue); broker behaves as expected but test fails; test_invalid_channel is safe enough for now.
  


git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@578053 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/framing/amqp_types.h
@@ -45,8 +45,8 @@ using std::string;
 typedef uint8_t FrameType;
 typedef uint16_t ChannelId;
 typedef uint32_t BatchOffset;
-typedef uint16_t ClassId;
-typedef uint16_t MethodId;
+typedef uint8_t ClassId;
+typedef uint8_t MethodId;
 typedef uint16_t ReplyCode;
 
 // Types represented by classes.
